requirements.txt
Flask==2.2.5
gunicorn==22.0.0
google-genai
